refacto

+++ libraries/Vasco_IOT/VDataLogger.h
@@ -26,22 +26,22 @@ class VDataLogger
   public:
 
     void start();
-    bool update(unsigned int delay);
+    bool update(int delay);
     
     void println(String text);
     void print(String text);
 
-    String dump(unsigned int delay);
+    String dump(int delay);
 
   private:
 
-    unsigned int _timer = 0;
+    long _timer = 0;
 
     void _addHistory(String text);
     
     struct message {
-      String        text;
-      unsigned int  time;
+      String  text;
+      long    time;
     };
     message _history[20];
 };
